+++ android/app/src/main/java/com/qwqlang465/ql/MainActivity.java
@@ -1,0 +1,5 @@
+package com.qwqlang465.ql;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}